+++ util.h
@@ -19,7 +19,7 @@ struct DataLoadingException : public std::exception {
 
 double euclid_distance(QPointF * point1, QPointF * point2);
 QRectF CenterRectToPoint(QRectF rect, QPointF point);
-QPointF PositionOnLine(Street street, int street_percentage);
+QPointF PositionOnLine(Street street, double street_percentage);
 QPen NextColor();
 
 #endif // UTIL_H
