+++ vehicle.h
@@ -9,7 +9,7 @@ public:
     Vehicle(int street_id, double speed, direction dir = true);
     ~Vehicle();
     int street_id;  // which street it's on
-    int street_percentage;
+    double street_percentage;
     direction dir;
     double speed;
 };
